requirements.txt
robotframework-browser
robotframework-seleniumlibrary
fastdeploy-python
pillow
imagehash
cryptography
datasets
pandas